Add files via upload
This project is a Machine Learning model that classifies emails as either spam or ham (not spam). The model is built using Logistic Regression, a popular algorithm for binary classification problems. The dataset used for this project is sourced from Kaggle and contains various emails.

The model has been trained and tested, yielding an accuracy of 94%. It demonstrates high precision and recall, particularly for classifying ham emails, where it achieved a precision of 0.88 and a recall of 1.00. For spam emails, the model achieved a precision of 1.00 and a recall of 0.91. These results indicate that the model is highly effective at correctly classifying both spam and ham emails.